requirements.txt
numpy>=1.21.2
esper>=1.5
PyQt6>=6.1.1
python-snap7>=1.1
opencv-python>=4.5.3.56
pyOpenGL>=3.1.5
pyGLM>=2.3.1